Cleanup: eliminate remaining sources of warnings with -Wpedantic

+++ src/libs/ck-libs/ampi/romio/mpi-io/file_c2f.c
@@ -37,7 +37,7 @@ Return Value:
 MPI_Fint MPI_File_c2f(MPI_File fh)
 {
 #ifndef INT_LT_POINTER
-    return (MPI_Fint) fh;
+    return (MPI_Fint)(intptr_t)fh;
 #else
     int i;
